Refactor and fix build
