+++ test-suite/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id("io.micronaut.application") version "3.7.7"
+    id("io.micronaut.minimal.application") version "3.7.7"
     id "io.micronaut.build.internal.jackson-xml-examples"
 }
 
@@ -14,6 +14,7 @@ dependencies {
     implementation("io.micronaut:micronaut-validation")
     implementation "io.projectreactor:reactor-core"
 }
+
 
 micronaut {
     runtime("netty")
