+++ test-suite/src/test/java/io/micronaut/xml/jackson/docs/Application.java
@@ -1,4 +1,4 @@
-package docs;
+package io.micronaut.xml.jackson.docs;
 
 import io.micronaut.runtime.Micronaut;
