+++ test-suite/src/test/java/io/micronaut/xml/jackson/docs/BookOperations.java
@@ -1,4 +1,4 @@
-package docs;
+package io.micronaut.xml.jackson.docs;
 
 import io.micronaut.core.async.annotation.SingleResult;
 import io.micronaut.http.MediaType;
